Move test data for dnlib platform to mock project
